DryFlowsheet

DryFlowsheet.CooledCrushingMill — Type

Represents a crushing device with cooling system.

Models

:TARGET_COOLANT_TEMP evaluates the heat transfer lost to coolant provided a target final stream temperature given by keyword argument temp_out. Product temperature is updated through an EnergyStream built with energy exchange computed through exchanged_heat, so that numerical value can be slightly different from target value.

:USING_GLOBAL_HTC makes use of a global heat transfer coefficient to evaluate heat flux across the cooling stream.

Attributes

rawmeal: The input meal applied to crushing process.

product: The output material stream at the end of product pipeline.

coolant: The output material stream at the end of cooling pipeline.

power: The power applied to the crushing process [W]

loss: The heat exchanged in between product and cooling pipelines [W].

globalhtc: Global heat transfer coefficient [W/K].

DryFlowsheet.EnergyStream — Type

Represents an energy stream.

Attributes

ḣ: Energy flow provided by stream [W].

DryFlowsheet.MaterialStream — Type

Represents a material stream.

Attributes

ṁ: Material mass flow rate [kg/s].

T: Stream temperature [K].

P: Stream pressure [Pa].

Y: Components mass fractions [-].

pipeline: Materials pipeline associated to Y.

DryFlowsheet.SolidsSeparator — Type

Represents a solids separator with efficiency η.

To-do's

Add inverse model to automatically tune efficiency η.

Attributes

η: Solids separation efficiency [-].

source: The stream to be separated into solids and others.

solids: The output solids stream.

others: The output remaining stream.

DryFlowsheet.StreamPipeline — Type

Array of materials to include in a stream.

DryFlowsheet.TransportPipeline — Type

Represents a pipeline with heat transfer.

Models

:TARGET_EXIT_TEMP evaluates the heat transfer lost to environment provided a target final stream temperature given by keyword argument temp_out. Product temperature is updated through an EnergyStream built with energy exchange computed through exchanged_heat, so that numerical value can be slightly different from target value.

:USING_GLOBAL_HTC makes use of a global heat transfer coefficient to evaluate heat flux across the pipe.

To-do's

Implement heat transfer losses through a convective heat transfer coefficient (HTC) computed from a suitable Nusselt number, for use of pipeline in simulation mode.

Attributes

product: The output material stream at the end of pipeline.

power: The heat exchanged in pipeline [W].

DryFlowsheet.cooled_crushing — Method

Manage use of CooledCrushingMill with different models.

DryFlowsheet.enthalpyflowrate — Function

Enthalpy flow rate of given stream [W].

DryFlowsheet.exchanged_heat — Method

Heat exchanged with stream to match outlet temperature.

DryFlowsheet.transport_pipe — Method

Manage use of TransportPipeline with different models.