DryCombustion

Empirical fuels

In industrial practice of CFD one is often confronted with simulating empirical fuels. This is how one generally calls a fuel provided in elemental mass fractions of elements and is the most common reporting format for heavy-fuel oil. Using EmpiricalFuel one can quickly perform conversions and find out the required air flow rate for setting up a process simulation or furnace operation.

DryCombustion.EmpiricalFuel — Type

Provides description of an empirical fuel based on elemental mass fractions.

elements::Vector{Symbol}: Chemical elements reported in fuel.

Y::Vector{Float64}: Array of provided masses per kilogram of fuel.

X::Vector{Float64}: Array of computed moles per kilogram of fuel.

Below we illustrate how to create a fuel with the approximate mass fractions of carbon and hydrogen in naphtalene; next we check its string representation.

julia> fuel = EmpiricalFuel([0.937, 0.063, 0.0]; scaler=:C=>10.0);

julia> fuel.X
3-element Vector{Float64}:
 10.0
  8.011606189967983
  0.0

julia> String(fuel)
"C(10.0000)H(8.0116)O(0.0000)"

Assuming the empirical fuel complete combustion can be represented by the following chemical equation:

\[1\:\mathrm{C}_x\mathrm{H}_y\mathrm{O}_z + a\:\mathrm{O}_2 + b\:\mathrm{N}_2 \rightarrow x\:\mathrm{CO}_2 + \dfrac{y}{2}\:\mathrm{H}_2\mathrm{O} + b\:\mathrm{N}_2\]

and since $\mathrm{CO}_2$ is the only product containing carbon and $\mathrm{H}_2\mathrm{O}$ the only with hydrogen, equilibrating the right-hand side of the equation is trivial; one can derive the value of $a$ that remains compatible with oxygen content in fuel as:

\[a = \dfrac{1}{2}\left(2x+\dfrac{y}{2}-z\right)\]

With this value it is trivial to find out the required mass flow rate of oxidizer. This is implemented in oxidizer_mass_flow_rate as documented below. Notice that this function will fail if one of the required elements in the above equation is missing. The presence of nitrogen oxides and sulphur is fuel is neglected in this calculations and one must assess whether they should be considered in a certain analysis prior to using this function.

DryCombustion.oxidizer_mass_flow_rate — Function

oxidizer_mass_flow_rate(f::EmpiricalFuel; y_o2 = 0.23)

Computes the required amount of oxidizer to perform complete combustion of 1 kg provided empirical fuel. The value of y_o2 represents the mass fraction of oxygen in oxidizer; default value is typical for air.

julia> fuel = EmpiricalFuel([0.937, 0.063, 0.0]; scaler=:C=>10);

julia> oxidizer_mass_flow_rate(fuel)
13.02691759229764

Heavy fuel-oils

Combustion of heavy-fuel oils (HFO) is discussed in detail by Lawn1987 [1]. Some relations that might be useful for the industrial combustion specialist are under implementation here to be integrated in larger models e.g. using DryFlowsheet, or simple calculations.

DryCombustion.hfo_specific_heat — Function

hfo_specific_heat(T::Float64, S::Float64)::Float64

Heavy fuel-oil specific heat estimation in terms of relative density $S$ as provided by Cragoe (1929). Temperature in kelvin.

DryCombustion.hfo_enthalpy_net_bs2869 — Function

hfo_enthalpy_net_bs2869(;
    ρ::Float64,
    x::Float64,
    y::Float64,
    s::Float64
)::Float64

Heavy fuel-oil net energy capacity accordinto to BS2869:1983. Value is computed in [MJ/kg]. Parameters are given as:

ρ: HFO density at 15 °C, [kg/m³].

water: Mass percentage of water, [%].

ash: Mass percentage of ashes, [%].

sulphur: Mass percentage of sulphur, [%].

julia> hfo_enthalpy_net_bs2869(; ρ = 1020.0, water = 0.1, ash = 0.05, sulphur = 1.0)
40.13509836320001

Discrete phase model

DryCombustion.fit_rosinrammler — Function

fit_rosinrammler(d₀, P₀; m=3.5)

Find parameter for particle size distribution with Weibull distribution, often called after Rosin-Rammler in the field of particles - based on characteristic size and associated cumulative density function (CDF) value. Parameter d₀ is the droplet size at which Weibull CDF evaluates to probability P₀. The value of m is generally recommended for a certain technology; a common value is provided by default.

DryCombustion.plot_rosinrammler — Function

Display Rosin-Rammler distribution and optionally reference data.