DryGranular

General porous media

Modeling of geometrical characteristics of porous beds is required for including both their thermal effect or role over chemistry in chemical reactors. A classical approach used in several commercial and open source tools is that of Gunn1978 [2]. In what follows we develop the ideas that lead to an analogous model which is implemented by this structure.

To build the model we will assume a reactor of constant rectangular cross-section ${A}_{r}={b}{w}$ and volume ${V}_{R}={b}{w}{h}$. Its cross-section perimeter is then ${P}_{R}=2({b}+{w})$. Inside this reactor we randomly pack cubic particles $\beta$ of surface area ${A}_{\beta}=6{l}_{\beta}^2$ and volume ${V}_{\beta}={l}_{\beta}^3$ at a porosity level ${\phi}$. Thus the total volume of solids inside the reactor is ${V}_{S}=(1-{\phi}){V}_{R}$ and the approximate number of particles ${N}=\frac{{V}_{S}}{{V}_{\beta}}$. Following a similar reasoning the total surface area of particles is ${A}_{S}={N}{A}_{\beta}$. Performing all the substitutions so far one finds the following expression

\[{A}_{S}=\frac{6(1-{\phi}){b}{w}{h}}{{l}_{\beta}}\]

Since the differential $d{A}={P}d{l}$ holds for the surface of a body over its length ${l}$, one can divide the above expression by the reactor length to get the perimeter of particles in a cross-section. We can further divide by the cross-section area itself and find the perimeter density which is a more general result, and find the expression proposed by Gunn1978 [2]. This result is summarized in the next equation where the subscript of particle size was dropped for generality.

\[{P} = \frac{6(1-{\phi})}{{l}}\]

An estimator of the number of channels per unit cross-section of reactor ${N}$ can be related to the porosity through ${N}\pi{R}^2={\phi}$. Because the above perimeter is shared between the fluid volume and solids, it holds that ${N}2\pi{R}=P$. Using these expressions one can solve for the porosity channels characteristic radius ${R}$ as given below, which is also a result reported by Gunn1978 [2].

\[{R}=\frac{{\phi}{l}}{3(1-{\phi})}\]

This model is probided in PackedBedPorosityDescriptor.

DryGranular.PackedBedPorosityDescriptor — Type

Provides description of porosity parameters with stochastic behavior.

ϕ::Union{Float64, Vector{Float64}}: Porosity volume fraction in medium [-].

l::Union{Float64, Vector{Float64}}: Characteristic particle size in medium [m].

σϕ::Union{Nothing, Float64}: Optional standard deviation of porosity volume fraction [-].

σl::Union{Nothing, Float64}: Optional standard deviation of characteristic particle size [m].

P::Union{Float64, Vector{Float64}}: Perimeter in reactor cross-section [m].

D::Union{Float64, Vector{Float64}}: Characteristic diameter of porosity channels [m].

A::Float64: Reactor area used for scaling perimeter [m²].

PackedBedPorosityDescriptor can be used to describe the geometry of exchange section of a packed bed for a single set of arguments.

julia> PackedBedPorosityDescriptor(; ϕ = 0.65, l = 0.10, area = 1.0)
PackedBedPorosityDescriptor(P = 21.000000 m, D = 0.123810 m)

It can also be used to describe randomly varying reactors, what is a more realistic thing to do when using this structure to simulate real world systems.

julia> PackedBedPorosityDescriptor(;
            ϕ  = 0.65, l  = 0.10,
            σϕ = 0.03, σl = 0.01,
            N = 2,
            ϕlims = (0.4, 0.8),
            llims = (0.0, 0.3),
            seed = 42,
            area = 1.0
        )
PackedBedPorosityDescriptor(
    P from  21.455749 m to  24.370742 m
    D from   0.125589 m to   0.102353 m
)

Rotary kiln models

In a rotary kiln as proposed by Kramers1952 [3]. Its goal is to be used as a process support tool or to integrate more complex models requiring integration of the bed profile. In its classical statement, the bed height profile $h(z)$ can be evaluated from volume of flowing material conservation through the following equations. Coordinate $z=0$ represents the discharge position where initial condition must be applied. This is given by the dam height, if any, or particle size.

\[\begin{aligned} \dfrac{dh}{dz} &= C₁ \left[\frac{h}{R}\left(2 - \frac{h}{R}\right)\right]^{-\frac{3}{2}} - C₂\\[6pt] C₁ &= \frac{3}{4}\dfrac{Φ\tan{γ}}{π R^3 ω}\\[6pt] C₂ &= \dfrac{\tan{β}}{\cos{γ}} \end{aligned}\]

The structure SymbolicLinearKramersModel implements the Kramers' ordinary differential equation for prediction of bed height profile in a rotary kiln. This equation is implemented under the formalism of ModelingToolkit.

DryGranular.SymbolicLinearKramersModel — Type

Creates a reusable linear Kramers model for rotary kiln simulation.

R::Symbolics.Num: Symbolic kiln internal radius

Φ::Symbolics.Num: Symbolic kiln feed rate

ω::Symbolics.Num: Symbolic kiln rotation rate

β::Symbolics.Num: Symbolic kiln slope

γ::Symbolics.Num: Symbolic solids repose angle

z::Symbolics.Num: Symbolic kiln axial coordinates

h::Symbolics.Num: Symbolic bed height profile

sys::ModelingToolkit.ODESystem: Problem ordinary differential equation

For integration of this model we implement RotaryKilnBedSolution. It provides the solved description of a rotary kiln bed geometry computed from the solution of bed height along the kiln length. The main goal of the quantities computed here is their use with heat and mass transfer models for the simulation of rotary kiln process. A simple post-processing utilitiy plotlinearkramersmodel is also provided.

DryGranular.RotaryKilnBedSolution — Type

General geometric description of a bed from Kramers equation solution.

z::Vector{Float64}: Solution coordinates [m]

h::Vector{Float64}: Solution bed height [m]

θ::Vector{Float64}: View angle from kiln center [rad]

l::Vector{Float64}: Bed-freeboard cord length [m]

A::Vector{Float64}: Local bed cross section area [m²]

η::Vector{Float64}: Local loading based on height [-]

ηₘ::Float64: Mean loading of kiln [%]

V::Float64: Bed integral volume [m³]

τ::Float64: Residence time of particles

β::Float64: Kiln slope [rad]

Arguments

Internal elements are initialized through the following constructor:

RotaryKilnBedSolution(z, h, β, R, Φ)

Where parameters are given as:

z: solution coordinates over length, [m].

h: bed profile solution over length, [m].

R: kiln internal radius, [m].

Φ: kiln feed rate, [m³/s].

An outer constructor is also provided for managing the integration of an instance of SymbolicLinearKramersModel. This is the recommended usage that is illustrated below.

Important: inputs must be provided in international system (SI) units as a better physical practice. The only exception is the rotation rate ω provided in revolution multiples. If the discharge end is held by a dam, its height must be provided instead of the particle size, as it is used as the ODE initial condition.

model: a symbolic kiln model.

L: kiln length, [m].

R: kiln internal radius, [m].

Φ: kiln feed rate, [m³/s].

ω: kiln rotation rate, [rev/s].

β: kiln slope, [rad].

γ: solids repose angle, [rad].

d: particle size or dam height, [m].

solver: Solver for DifferentialEquations. Defaults to Tsit5.

rtol: Relative integration tolerance. Defaults to 1.0e-08.

atol: Absolute integration tolerance. Defaults to 1.0e-08.

DryGranular.plotlinearkramersmodel — Function

plotlinearkramersmodel(
    model::RotaryKilnBedSolution;
    normz::Bool = false,
    normh::Bool = false
)::Figure

Standardized plotting of RotaryKilnBedSolution bed profile. It supports normalization of axes throught keywords normz for axial coordinate and normh for bed depth.

Data in next example is an SI conversion of an example from Kramers1952 [3].

julia> L = 13.715999999999998;  # Kiln length [m]

julia> D = 1.8897599999999999;  # Kiln diameter [m]

julia> β = 2.3859440303888126;  # Kiln slope [°]

julia> γ = 45.0;                # Repose angle [°]

julia> d = 1.0;                 # Particle/dam size [mm]

julia> Φ = 10.363965852671996;  # Feed rate [m³/h]

julia> ω = 3.0300000000000002;  # Rotation rate [rev/min]

julia> bed = RotaryKilnBedSolution(;
            model = SymbolicLinearKramersModel(),
            L     = L,
            R     = D / 2.0,
            Φ     = Φ / 3600.0,
            ω     = ω / 60.0,
            β     = deg2rad(β),
            γ     = deg2rad(γ),
            d     = d / 1000.0
        );

julia> bed
RotaryKilnBedSolution(τ = 13.169938 min, ηₘ = 5.913271 %)

julia> bed.τ
790.1963002204403

In the following dummy example we force a very thick analytical bed solution, filling the radius of the rotary drum.

julia> R = 1.0e+00;

julia> Φ = 1.0e-02;

julia> z = collect(0.0:0.1:10.0);

julia> h = R * ones(size(z));

julia> Aₐ = π * R^2 / 2;

julia> Vₐ = Aₐ * z[end];

julia> bed = RotaryKilnBedSolution(z, h, 0, R, Φ)
RotaryKilnBedSolution(τ = 26.179939 min, ηₘ = 50.000000 %)

Next we confirm the internal evaluations of the model match the expected analytical values.

julia> mean(bed.θ) ≈ π
true

julia> mean(bed.l) ≈ 2R
true

julia> mean(bed.A) ≈ Aₐ
true

julia> mean(bed.η) ≈ 0.5
true

julia> bed.ηₘ ≈ 50.0
true

julia> bed.V ≈ Vₐ
true

julia> bed.τ ≈ Vₐ / Φ
true

Validation of Kramers' model is provided here.

Finally a set of basic equations provided for process analysis.

DryGranular.sullivansηₘ — Function

Sullivans approximation to kiln filling.

DryGranular.dimlessNΦ — Function

Kramers (1952) dimensionless group NΦ.

DryGranular.dimlessNₖ — Function

Kramers (1952) dimensionless group Nₖ.

DryGranular.perrayresidence — Function

Compute residence time from Peray's equation.

DryGranular.kramersnlapprox — Function

Nonlinear formulation of Kramers model approximate solution.