DryMaterials

Solid Thermodynamics

Danger

The following sub-sections document structures and functions that are undergoing active development to harmonize all interfaces. Changes are expected to happen all the time before a common ground is found.

Laurent Polynomial Models

DryMaterials.SolidPolynomialProperties — Type

Arbitrary Laurent polynomial specific heat and enthalpy of materials.

c: Polynomial representation of specific heat [J/(kg.K)].

h: Polynomial representation of enthalpy [J/kg].

hₒ: Reference state enthalpy [J/kg].

Shomate Model

Solid materials properties are often reported according to the formalism of Shomate1954 [4]. To be able to handle data under this format, the following structure is provided.

DryMaterials.SolidShomateProperties — Type

Thermodynamic properties of solids with Shomate functions.

IMPORTANT: the implementation of makestepwise1d used for step-wise function evaluation takes the mean of both ranges at the T_ch, while the actual Shomate uses the high range for doing so. The function enforces the right behavior by multiplying T_ch the value by 1-eps().

cₚ: Molar specific heat [J/(mol.K)].

h: Molar enthalpy [J/mol].

s: Molar entropy [J/K].

a_lo: Low temperature range Shomate coefficients.

a_hi: High temperature range Shomate coefficients.

T_ch: Temperature of range change for evaluation.

Using data from NIST Webbook of Chemistry we have the test:

julia> T = [298.0, 300.0, 400.0, 847.0, 900.0, 1900.0];
    
julia> c = [44.57, 44.77, 53.43, 67.42, 67.95, 77.99];

julia> silica = DryMaterials.getshomatesilica();

julia> sum(abs2, silica.cₚ.(T) - c) < 0.0001
true

The inner workings of DryMaterials.SolidShomateProperties are provided by the following functions. Notice that these do not automatically manage the ranges, so generating functions with the main structure is the recommended way of implementing programs.

DryMaterials.shomatespecificheat — Function

shomatespecificheat(T::Float64, c::Vector{Float64})::Float64

Molar specific heat with Shomate equation [J/(mol.K)].

DryMaterials.shomateenthalpy — Function

shomateenthalpy(T::Float64, c::Vector{Float64})::Float64

Molar enthalpy with Shomate equation [J/mol].

DryMaterials.shomateentropy — Function

shomateentropy(T::Float64, c::Vector{Float64})::Float64

Entropy with Shomate equation [J/K].

Ongoing documentation

DryMaterials.SolidMineralPhase — Type

Represents a solid mineral phase for thermodynamic calculations.

name: Name of phase.

ρ: Specific mass of phase [kg/m³].

M: Molecular mass of phase [kg/mol].

p: Polynomial properties of phase.

DryMaterials.enthalpy — Function

Evaluates the enthalpy of material [J/kg].

DryMaterials.specificheat — Function

Evaluates the specific heat of materials [J/(kg.K)].

DryMaterials.density — Function

Evaluates the density of material [kg/m³].

DryMaterials.molecularmass — Function

Retrieve molecular mass of materials [kg/mol].

Hard-coded materials

DryMaterials.H_AIR — Constant

Coefficients for air enthalpy polynomial [J/kg].

DryMaterials.PureAir — Type

Simple implementation of pure air for illustration purposes.

DryMaterials.PureMineral — Type

Simple solid mineral material for illustration purposes.

DryMaterials.PureWater — Type

Simple implementation of liquid water for illustration purposes.

Temporary (moving to database)

DryMaterials.getshomatesilica — Function

Representation of crystaline silica $SiO_2$ (quartz).

TODO this contains only the thermodynamics, not a SolidMineralPhase!

DryMaterials.getamorphoussilica — Function

Representation of amorphous silica $SiO_2(a)$.

TODO recover missing defaults in Schieltz (1964).

DryMaterials.getkaolinite — Function

Representation of Kaolinite $Al_2Si_2O_5(OH)_4$.

TODO recover missing defaults in Schieltz (1964).

DryMaterials.getmetakaolin — Function

Representation of metakaolin $Al_2Si_2O_7$.

TODO recover missing defaults in Schieltz (1964).

DryMaterials.getspinel — Function

Representation of spinel $Al_4Si_3O_{12}$.

TODO recover missing defaults (somewhere else, not in Schieltz (1964)).