DryTransport

Dimensionless groups

Groups by definition

DryTransport.nusselt — Function

Evaluate an arbitrary Nusselt number relationship.

DryTransport.prandtl — Function

Evaluate Prandtl number relationship.

DryTransport.reynolds — Function

Evaluate Reynolds number relationship.

Groups by application

DryTransport.ConstantPrandtl — Type

Prandtl number set to a constant value.

DryTransport.ReynoldsPipeFlow — Type

Reynolds number applied to internal pipe flow.

DryTransport.NusseltGnielinski — Type

Gnielinski's relationship for Nusselt number.

DryTransport.NusseltDittusBoelter — Type

Dittus-Boelter's relationship for Nusselt number.

Heat transfer coefficients

DryTransport.HtcPipeFlow — Type

Heat transfer coefficient for internal pipe flow.

DryTransport.htc — Function

Implements the interface for heat transfer coefficient evaluation.

Notice that although temperature is provided in this interface, it is used only for Pr calculation. Other properties might have arbitrarily complex dependencies and types, so it was chosen by design to keep their evaluation to the user before calling this.

Also, to ensure internal consistency because of Nusselt number dependency on Prandtl number, thermal conductivity is evaluated from specific heat and viscosity. It is up to the user to make sure the provided Prandtl number is compatible.