Documents

The following functionalities are intended mostly for the own toolbox maintenance and documentation generation. This can also be imported as a replacement for some boilerplate code in other projects. The use of the functions provided in this module is essentially illustrated in docs/make.jl.

WallyToolbox.Documents.deployrepo — Function

Generate a repo argument in the format expected by deploydocs.

WallyToolbox.Documents.setdocmetawrapper! — Function

Simple wrapper to set documentation tests metadata.

WallyToolbox.Documents.get_format — Function

Get format specified to generate docs.

Because most of my work is done with Obsidian and the tool does not support Julia math blocks and a few other features, the functions below were conceived to help documentation conversion without modifications to standard $\LaTeX$ syntacs in Obsidian.

WallyToolbox.Documents.julianizemarkdown — Function

Convert equations from dollar to Julia ticks notation.

WallyToolbox.Documents.formatequations — Function

Convert (multiline) equations to Julia markdown.

WallyToolbox.Documents.formatnotecells — Function

Convert cells of notebooks to Documenter format.

There are some media, e.g. external plots, one might want to leave out of formatting workflow; the next functions use the alternative text field of those to make sure they must be included in the pre-processing. For marking an image or video for standardized processing one add a @image or @video to them.

WallyToolbox.Documents.formatembimages — Function

Convert simple embeded markdown image to raw HTML.

WallyToolbox.Documents.formatembvideos — Function

Convert simple embeded markdown video to raw HTML.

Finally, we wrap up a standard workflow for converting Pluto notebooks.

WallyToolbox.Documents.convert_pluto — Function

Convert Pluto notebook into a static HTML file.