Constants

Configuration

The following constants are not exported by the toolbox; their use is mostly intended for developers and some particular use cases. If extending module data, please consider using the reference path.

WallyToolbox.WALLYTOOLBOXPATH — Constant

Path to supporting modules annex to WallyToolbox.

WallyToolbox.WALLYTOOLBOXDATA — Constant

Path to data files used by WallyToolbox.

WallyToolbox.MAJORDOMEPATH — Constant

Path to Python modules from legacy Majordome package.

Physical constants

WallyToolbox.GAS_CONSTANT — Constant

Ideal gas constant [8.31446261815324 [J/(mol.K)].

WallyToolbox.STEFAN_BOLTZMANN — Constant

Stefan-Boltzmann constant [5.670374419e-8 W/(m².K⁴)].

Reference states

WallyToolbox.T_REF — Constant

Normal atmospheric temperature [273.15 K].

WallyToolbox.P_REF — Constant

Reference atmospheric pressure [101325.0 Pa].

WallyToolbox.C_REF — Constant

Normal state concentration [44.61503340547032 mol/m³].

Unit conversion

WallyToolbox.ZERO_CELSIUS — Constant

Zero degrees Celsius in Kelvin [273.15 K].

WallyToolbox.ONE_ATM — Constant

Atmospheric pressure at sea level [101325.0 Pa].

WallyToolbox.JOULE_PER_CALORIE — Constant

Conversion factor from calories to joules [4.184 J/cal].

Chemistry-related

WallyToolbox.M_AIR — Constant

Air mean molecular mass [0.0289647 kg/mol].

WallyToolbox.ELEMENTS — Constant

Periodic table of stable chemical elements.