Flowsheet

Unit operations

WallyToolbox.StreamPipeline — Type

Array of materials to include in a stream.

WallyToolbox.MaterialStream — Type

Represents a material stream.

Fields

ṁ::Float64: Material mass flow rate [kg/s].

T::Float64: Stream temperature [K].

P::Float64: Stream pressure [Pa].

Y::Vector{Float64}: Components mass fractions [-].

pipeline::StreamPipeline: Materials pipeline associated to Y.

WallyToolbox.EnergyStream — Type

Represents an energy stream.

Fields

ḣ::Float64: Energy flow provided by stream [W].

WallyToolbox.TransportPipeline — Type

Represents a pipeline with heat transfer.

Models

:TARGET_EXIT_TEMP evaluates the heat transfer lost to environment provided a target final stream temperature given by keyword argument temp_out. Product temperature is updated through an EnergyStream built with energy exchange computed through exchanged_heat, so that numerical value can be slightly different from target value.

:USING_GLOBAL_HTC makes use of a global heat transfer coefficient to evaluate heat flux across the pipe.

To-do

Implement heat transfer losses through a convective heat transfer coefficient (HTC) computed from a suitable Nusselt number, for use of pipeline in simulation mode.

Fields

product::MaterialStream: The output material stream at the end of pipeline.

power::EnergyStream: The heat exchanged in pipeline [W].

WallyToolbox.SolidsSeparator — Type

Represents a solids separator with efficiency η.

To-do

Add inverse model to automatically tune efficiency η.

Fields

η::Float64: Solids separation efficiency [-].

source::MaterialStream: The stream to be separated into solids and others.

solids::MaterialStream: The output solids stream.

others::MaterialStream: The output remaining stream.

WallyToolbox.CooledCrushingMill — Type

Represents a crushing device with cooling system.

Models

:TARGET_COOLANT_TEMP evaluates the heat transfer lost to coolant provided a target final stream temperature given by keyword argument temp_out. Product temperature is updated through an EnergyStream built with energy exchange computed through exchanged_heat, so that numerical value can be slightly different from target value.

:USING_GLOBAL_HTC makes use of a global heat transfer coefficient to evaluate heat flux across the cooling stream.

Fields

rawmeal::MaterialStream: The input meal applied to crushing process.

product::MaterialStream: The output material stream at the end of product pipeline.

coolant::MaterialStream: The output material stream at the end of cooling pipeline.

power::EnergyStream: The power applied to the crushing process [W]

loss::EnergyStream: The heat exchanged in between product and cooling pipelines [W].

globalhtc::Union{Nothing, Float64}: Global heat transfer coefficient [W/K].

WallyToolbox.transport_pipe — Function

Manage use of TransportPipeline with different models.

WallyToolbox.cooled_crushing — Function

Manage use of CooledCrushingMill with different models.

Computed quantities

WallyToolbox.enthalpyflowrate — Function

Enthalpy flow rate of given stream [W].

WallyToolbox.exchanged_heat — Function

Heat exchanged with stream to match outlet temperature.