Helpers

The core packages of WallyToolbox ecosystem provide shared functionalities and types that are used in several other more specialized packages. This allows for standardization of interfaces, employed quantities, and avoid boilerplate code. This page organizes everything to facilitate the understanding of the end-user. Knowing the base helper packages DryConstants and DryUtilities is key for proper usage of WallyToolbox.

DryConstants

Physical constants

DryConstants.GAS_CONSTANT — Constant

Ideal gas constant [J/(mol.K)].

DryConstants.STEFAN_BOLTZMANN — Constant

Stefan-Boltzmann constant [W/(m².K⁴)].

Reference states

DryConstants.T_REF — Constant

Normal atmospheric temperature [K].

DryConstants.P_REF — Constant

Reference atmospheric pressure [Pa].

DryConstants.C_REF — Constant

Normal state concentration [mol/m³].

Other constants

DryConstants.ZERO_CELSIUS — Constant

Zero degrees Celsius in Kelvin [273.15 $K$].

DryConstants.ONE_ATM — Constant

Atmospheric pressure at sea level [101325.0 $Pa$].

DryConstants.M_AIR — Constant

Air mean molecular mass [kg/mol].

DryUtilities

Haskell-like array slicing

Those who know Haskell probably started learning it by manipulating lists with head and tail. Those functionalities are not available in Julia by default and array slicing - with an ugly syntax - is required. Since this is done often in the fields of application of DryTooling, both head and tail together with a body functions are available in its core. They are simple wrapers over the @view macro and work with both iterable types and arrays. The following snippet illustrates their usage.

julia> v = collect(1:4);

julia> head(v) == [1; 2; 3]
true

julia> tail(v) == [2; 3; 4]
true

julia> body(v) == [2; 3]
true

More examples are provided in the following documentation ahead of each of the functions.

julia> head(1:4)
1:3

julia> head([1, 2, 3, 4])
3-element view(::Vector{Int64}, 1:3) with eltype Int64:
 1
 2
 3

DryUtilities.head — Function

head(z)

Access view of array head. See also tail and body.

julia> tail([1, 2, 3, 4])
3-element view(::Vector{Int64}, 2:4) with eltype Int64:
 2
 3
 4
julia> tail(1:4)
2:4

DryUtilities.tail — Function

tail(z)

Access view of array tail. See also head and body.

julia> body([1, 2, 3, 4])
2-element view(::Vector{Int64}, 2:3) with eltype Int64:
 2
 3
julia> body(1:4)
2:3

DryUtilities.body — Function

body(z)

Access view of array body. See also head and tail.

Handling of discontinuous functions

Discontinuous functions are all over in real world applications. Whether they handle discrete signals sent to controllers or represent a material property change in the solution domain of a heat transfer simulation, they are often represented by a single or a composition of Heaviside step functions. Again, because its implementation is pretty simple and optimization routines require a differentiable form of this function, DryUtilities implements heaviside and interval as proposed in this StackOverflow answer.

julia> heaviside(-1) == 0
true

julia> heaviside(-1.0) == 0.0
true

julia> heaviside(0.0) == 0.5
true

julia> heaviside(1.0) == 1.0
true

julia> interval(10; a = 0, b = 10) == 0.5
true

We see below that heaviside also works on ranges

julia> heaviside(-2:2)
5-element Vector{Float64}:
 0.0
 0.0
 0.5
 1.0
 1.0

DryUtilities.heaviside — Function

heaviside(t)

Provides a Heaviside function compatible with automatic differentiation. This is a requirement for conceiving, e.g., model predictive controls with discontinuous functions under ModelingToolkit.

By implementation inheritance that is also the case for interval:

julia> interval(0:6; a = 2, b = 5)
7-element Vector{Float64}:
 0.0
 0.0
 0.5
 1.0
 1.0
 0.5
 0.0

DryUtilities.interval — Function

interval(x; a=-Inf, b=Inf)

Returns 1 if $x ∈ (a, b)$, 1/2 for $x = a || x = b$, or 0.

As it is the case for representation of specific heats using NASA7/NASA9 or Shomate polynomials, functions defined by parts with an specific change point are also required in physical modeling. To this end, a stepwise function can be established with makestepwise1d. If keyword differentialble = true, then the function makes use of the above interval and remains compatible with ModelingToolkit, for instance.

julia> f = makestepwise1d(x->x, x->x^2, 1.0; differentiable = true);

julia> f(0:0.2:2.0)
11-element Vector{Float64}:
 0.0
 0.2
 0.4
 0.6
 0.8
 1.0
 1.44
 1.9599999999999997
 2.5600000000000005
 3.24
 4.0

julia> using ModelingToolkit

julia> @variables x
1-element Vector{Num}:
 x

julia> f(x); # Output is too long, try by yourself.

DryUtilities.makestepwise1d — Function

makestepwise1d(lo, hi, xc)

Creates an univariate function that is composed of two parts, the first evaluated before a critical domain point xc, and the second above that value. This is often required, for instance, for the evaluation of NASA polynomials for thermodynamic properties. If differentiable, then the returned function is compatible with symbolic argument as required when using package ModelingToolkit, etc.

Rounding numbers and automatic axes

Danger

This section documents functions that are used in a very unstable context.

Simple rounding is not enough. Getting values that are rounded close to a power of a given number and rounded to floor or ceil is often the case. This is standardized in DryTooling through closestpowerofx:

julia> closestpowerofx(10)
10

julia> closestpowerofx(11)
20

julia> closestpowerofx(11, roundf = floor)
10

julia> closestpowerofx(11, x = 5, roundf = floor)
10

julia> closestpowerofx(12.0; x = 10)
20

julia> closestpowerofx(12.0; x = 10, roundf = floor)
10

julia> closestpowerofx(12.0; x = 10, roundf = round)
10

DryUtilities.closestpowerofx — Function

closestpowerofx(
    v::Number;
    x::Number = 10,
    roundf::Function = ceil
)::Int64

Compute the integer power of x closest to v using roundf as rouding method. This might be useful for automatic setting more reasonable limits to plot axis or similar applications. Changing the rouding method through roundf is also possible.

Below we illustrate the usage of axesunitscaler.

NOTE: this function is not yet stable. In the future it will instead return labels using symbols like k, M, G, etc., for the units through a flag provided by the user.

julia> axesunitscaler(1)
("", 1)

julia> axesunitscaler(1000)
("[×1000]", 1000)

julia> axesunitscaler(1000000)
("[×1000000]", 1000000)

DryUtilities.axesunitscaler — Function

axesunitscaler(x::Number)::Tuple{String, Int64}

Find scaling factor for multiples of 1000 units. Together with closestpowerofx this can be used to produce better automatic plot axes limits. The returned values provide the string for modifying the axis label and the associated scaling factor.

Computation of changes and residuals

Danger

This section documents functions that are used in a very unstable context.

DryUtilities.maxabsolutechange — Function

Maximum absolute change in a solution array.

DryUtilities.maxrelativechange — Function

Maximum relative change in a solution array.

Unit conversion

DryUtilities.nm3_h_to_kg_h — Function

Convert [Nm³/h] to [kg/h].

DryUtilities.kg_h_to_nm3_h — Function

Convert [kg/h] to [Nm³/h].

Others

DryUtilities.defaultvalue — Function

Syntax sugar for handling a possibly nothing value.

Abstract types

Fixed state materials

DryAbstract.AbstractMaterial — Type

Base type for any material substance.

DryAbstract.AbstractSolidMaterial — Type

Base type for solid materials.

DryAbstract.AbstractLiquidMaterial — Type

Base type for liquid materials.

DryAbstract.AbstractGasMaterial — Type

Base type for gas materials.

DryAbstract.AbstractSolidMineralPhase — Type

Base type for any mineral material.

Problem solving and physical models

DryAbstract.AbstractMatrixProblem — Type

Base type for linear algebra problems.

DryAbstract.AbstractIterativeSolver — Type

Base type for (nonlinear) iterative solvers.

DryAbstract.AbstractSolutionStorage — Type

Base type for storing simulation solution.

DryAbstract.AbstractPhysicalModel — Type

Base type for physical models.

Transport, thermodynamics, and kinetics

DryAbstract.AbstractTransportModel — Type

Base type for transport models.

DryAbstract.AbstractSolidTransport — Type

Base type for transport models.

DryAbstract.AbstractGasThermo — Type

Base type for thermodynamic models.

DryAbstract.AbstractSolidThermo — Type

Base type for thermodynamic models.

DryAbstract.AbstractMixtureSubstance — Type

Base type for simplified mixture substances.

DryAbstract.AbstractMixturePhase — Type

Base type for simplified mixture phases.

DryAbstract.AbstractKineticsMechanism — Type

Base type for coded kinetics mechanisms.

Finite volume method and relatives

DryAbstract.AbstractDiffusionModel1D — Type

Base type for diffusion (heat, species, ...) models.

DryAbstract.AbstractGrid1D — Type

Base type of one-dimensional grids.