DryTransport

Dimensionless groups

When dealing with fluid flow and reactor models, it is always useful to be able to quickly compute approximate dimensionless numbers for the studied case. In what follows we provide a brief description of some of these quantities and some context for their use whenever possible. Definitions might vary according to the author or field of application and here we follow mostly Bird2001 [5].

Although quite unusual, it seems it is time to start this list with the Knudsen number, which evaluates the particles mean free path over system characteristic dimension. Division between rarefied gas (Boltzmann) and continuum mechanics (Navier-Stokes).

Although an implementation is not yet provided here, users are encouraged to estimate its value based on some mean free path estimation for their system before following with the analysis of a transport problem. Generally speaking, if continuum mechanics hypothesis is not valid, i.e Knudsen number is too high, most of what follows cannot be applied, thus justifying why this group comes first.

Common dimensionless groups

Reynolds dimensionless group is named for Osborne Reynolds (1842-1912), professor of engineering at the University of Manchester. He studied the laminar-turbulent transition, turbulent heat transfer, and theory of lubrication Bird2001 [5]. In general we denote Reynolds number by $\mathrm{Re}$ and it is used to delineate flow regimes. For circular tubes it is defined as:

\[\mathrm{Re} = \frac{\rho \langle v_{z} \rangle D}{\mu}\]

where $\langle{}v_{z}\rangle$ is the average flow velocity in axial direction and $D$ is the tube diameter. For values up 2100 the flow is assumed laminar if steady state is established and density is constant. For more, see Bird2001 [5], Chapter 2.

DryTransport.reynolds — Function

Evaluate Reynolds number relationship.

The Nusselt number provides the ratio of convective to conductive heat transfer at a boundary in a fluid, defined as

\[\mathrm{Nu}=\frac{hL}{k}\]

Often in buoyancy-driven flow analysis it is correlated as $\mathrm{Nu}=a\mathrm{Ra}^b$. A Nusselt number of value one represents heat transfer by pure conduction. Increasing this number implies a laminar conductive-dominant flow and then a convective dominant turbulent flow.

DryTransport.nusselt — Function

Evaluate Nusselt number relationship.

Prandtl represents the ratio of momentum diffusivity to thermal diffusivity $\mathrm{Pr}=\frac{\nu}{\alpha}$. High $\mathrm{Pr}$ indicates that momentum transfer is more effective than heat transfer (oils), while low values (liquid metals) indicate thermal boundary layer is more important than viscous one.

Ludwig Prandtl (1875-1953) (pronounced "Prahn-t'), who taught in Hannover and Gottingen and later served as the Director of the Kaiser Wilhelm Institute for Fluid Dynamics, was one of the people who shaped the future of his field at the beginning of the twentieth century; he made contributions to turbulent flow and heat transfer, but his development of the boundary-layer equations was his crowning achievement Bird2001 [5]. The dimensionless quantity appear under two forms of interest for the analysis of reactors: its thermal and its chemical versions. In thermal version, this number compares the kinematic viscosity $\nu$ to the thermal diffusivity $\alpha$, which is replaced by species diffusivity in its chemical version, which is more often referred to as Schmidt number. Ernst Heinrich Wilhelm Schmidt (1892-1975), who taught at the universities in Gdansk, Braunschweig, and Munich (where he was the successor to Nusselt) Bird2001 [5]_. The ratio $\frac{\nu}{\alpha}$ indicates the relative ease of momentum and energy or species transport in flow systems. This dimensionless ratio in thermal form is given by

\[\mathrm{Pr} = \frac{\nu}{\alpha} = \frac{C_{p} \mu}{k}\]

If transport properties for a gas are not available, thermal Prandtl number can be estimated at low pressure and non-polar molecules mixtures with help of Eucken formula as

\[\mathrm{Pr} = \frac{C_{p}}{C_{p} + \frac{5}{4}R}\]

DryTransport.prandtl — Function

Evaluate Prandtl number relationship.

Péclet Jean-Claude-Eugene Peclet (pronounced "Pay-clay" with the second syllable accented) (1793-1857) authored several books including one on heat conduction Bird2001 [5]. This number is nothing more than the multiplication of Reynolds and Prandtl or Schmidt numbers. By simplifying factors one easily determines that it represents the ratio of convective by diffusive transport (thermal or species). High $\mathrm{Pe}$ limit represents the #plug-flow behavior.

\[\mathrm{Pe}_{th} = \mathrm{Re} \mathrm{Pr}\qquad \mathrm{Pe}_{ch} = \mathrm{Re} \mathrm{Sc}\]

Convection-related groups

Grashof number named after Franz Grashof (1826-1893) (pronounced "Grahss-hoff). He was professor of applied mechanics in Karlsruhe and one of the founders of the Verein Deutscher Ingenieure in 1856 Bird2001 [5]. The Grashof number is the characteristic group occurring in analyses of free convection. It approximates the ratio of the buoyancy to viscous force acting on a fluid, defined as

\[\mathrm{Gr}=\frac{g\beta(T_s-T_{\infty})L^3}{\nu^2}\]

and is analogous to Reynolds number in natural convection. Increasing the value of this number above a given threshold promotes buoyancy driven flow.

DryTransport.grashof — Function

Evaluate Grashof number relationship.

Rayleigh number is the product of Grashof $\mathrm{Gr}$ and Prandtl $\mathrm{Pr}$ numbers. Related to the transition from laminar to turbulent in buoyancy-driven flows. Laminar to turbulent is assumed to take place at $10^9$ Balaji2014 [6].

Mass transfer groups

Schmidt number is the mass diffusion equivalent of Prandtl's. Its range can be much broader than its thermal relative, Prandtl number. This is given by the effects of cross-section and molar weight determining mass diffusivity of gas species. For more, see Bird2001 [5], Chapter 9.

\[\mathrm{Sc} = \frac{\nu}{D}\]

Sherwood number, also called the mass transfer Nusselt number is a dimensionless number used in mass-transfer operation. It represents the ratio of the total mass transfer rate (convection + diffusion) to the rate of diffusive mass transport, and is named in honor of Thomas Kilgore Sherwood.

Multiphase-specific

Weber group is often found in applications of multiphase flows where strongly curved surfaces are present. It represents the ratio of drag forces to cohesion forces, and can be thought of as a measure of the relative importance of the fluid's inertia compared to its surface tension. As reminded by Amsden1989 [7], for $\mathrm{We}>1$, drop oscillations, distortions, and breakup must be considered, requiring other sub-models other than simple drag to describe the flow.

Groups by application

DryTransport.ConstantPrandtl — Type

Prandtl number set to a constant value.

DryTransport.ReynoldsPipeFlow — Type

Reynolds number applied to internal pipe flow.

DryTransport.NusseltGnielinski — Type

Gnielinski's relationship for Nusselt number.

DryTransport.NusseltDittusBoelter — Type

Dittus-Boelter's relationship for Nusselt number.

Heat transfer coefficients

DryTransport.HtcPipeFlow — Type

Heat transfer coefficient for internal pipe flow.

DryTransport.htc — Function

Implements the interface for heat transfer coefficient evaluation.

Notice that although temperature is provided in this interface, it is used only for Pr calculation. Other properties might have arbitrarily complex dependencies and types, so it was chosen by design to keep their evaluation to the user before calling this.

Also, to ensure internal consistency because of Nusselt number dependency on Prandtl number, thermal conductivity is evaluated from specific heat and viscosity. It is up to the user to make sure the provided Prandtl number is compatible.

Property models

Medium properties often are non-constant and require description through different sorts of models for representing their dependence on solution quantities, such as temperature, pressure, composition, etc. This section is devoted to document such models.

Polynomial properties

The most commonly used representation of thermal conductivity of materials is through polynomial fits of temperature. Although this approach does not provide any physical-based representation, it is easy to use and fast to evaluate in most computational science problems. A common interface for polynomial properties is given by the following structure.

DryTransport.TempPolynomialHeatConductivity — Type

Wrapper for a polynomial temperature-dependent heat conductivity.

p::Polynomials.Polynomial: Heat conductivity polynomial.

The general use case of this is to create objects compatible with the function object approach employed for properties evaluation across the module.

julia> k = TempPolynomialHeatConductivity([1.5, -0.001])
TempPolynomialHeatConductivity(Polynomial(1.5 - 0.001*T))

julia> k(1000.0)
0.5

Although not the most efficienty way, a simple wrapper for providing constant heat conductivity remaining compatible with other funcionalities is provided:

julia> k = constheatconductivity(5.0)
TempPolynomialHeatConductivity(Polynomial(5.0))

DryTransport.constheatconductivity — Function

Constant heat conductivity wrapper compatible with temperature dependency.

An analogous interface is also provided for viscosity temperature dependence.

DryTransport.TempPolynomialFluidViscosity — Type

Wrapper for a polynomial temperature-dependent fluid viscosity.

p::Polynomials.Polynomial: Fluid viscosity polynomial.

Usage is analogous to TempPolynomialHeatConductivity. ```

DryTransport.constfluidviscosity — Function

Constant heat conductivity wrapper compatible with temperature dependency.

Granular media

According to Hanein2017 [8] the representation of effective thermal conductivity of a solids bed in a rotary kiln can be approximated through a Maxell model based on effective medium theory. To keep track of eventually temperature-dependent properties and make use of this model, the following interfaces are provided.

DryTransport.GranularMediumHeatConductivity — Type

Provides the heat conductivity of a solids granular medium embeded in gas.

ks::WallyToolbox.AbstractHeatCondTemperatureDep: Heat conductivity model for solid phase.

kg::WallyToolbox.AbstractHeatCondTemperatureDep: Heat conductivity model for gas phase.

ϕ::Float64: Solids packing fraction.

This composite type relies on a gas and a solid; below we illustrate how to evaluate a granular medium effective heat conductivity using this structure.

julia> ks = constheatconductivity(5.0);

julia> kg = constheatconductivity(0.092);

julia> kb = GranularMediumHeatConductivity(ks, kg, 0.36);

julia> kb(300.0)
0.23471049304677621

DryTransport.maxwell_eff_conductivity — Function

maxwell_eff_conductivity(kg, ks, ϕ)

Maxwell effective medium theory of thermal conductivity computed in terms of gas thermal conductivity kg, solids thermal conductivity ks, and solids packing fraction ϕ.

Air properties

For the simulation of rotary kilns, Mujumdar2006i [9] proposes some data for air properties implemented by the following interfaces. It must be noted that the thermal conductivity proposed by the authors quickly diverges above 1500 K and users must be aware of its implications.

DryTransport.AirHeatConductivityMujumdar2006 — Type

Air heat conductivity for rotary kiln simulation - Mujumdar (2006).

julia> k = AirHeatConductivityMujumdar2006();

julia> k(300)
0.027600315300000004

DryTransport.AirViscosityMujumdar2006 — Type

Air viscosity for rotary kiln simulation - Mujumdar (2006).

julia> μ = AirViscosityMujumdar2006();

julia> μ(300)
1.837988950255163e-5