OpenFOAM

OpenFOAM.OPENFOAMBANNER — Constant

Banner for all OpenFOAM files with right version.

OpenFOAM.AbstractFlowRateProfile — Type

Abstract type for flow rate profile distributions.

OpenFOAM.AbstractInjectionModel — Type

Abstract type for injection models.

OpenFOAM.AbstractSizeDistribution — Type

Abstract type for particle size distributions.

OpenFOAM.ConstantFlowRateProfile — Type

Constant flow rate profile.

OpenFOAM.InjectionModel — Type

Represents a generic particle injection model.

Constraints imposed by OpenFOAM:

nParticle makes massFlowRate and massTotal to be ignored.

massTotal cannot be used in steady state simulations.

if flowRateProfile is not found a constant massFlowRate profile is computed from the ratio of massTotal and duration.

To-do:

Check if parcelBasisType from is applicable to derived types.

massFlowRate::Number: Particles mass flow rate if nParticle is not provided.

massTotal::Number: Total mass to be injected if nParticle is not provided.

SOI::Number: Start of injection in seconds (base).

duration::Number: Duration of injection in seconds (base).

nParticle::Number: Number of particles per parcel.

uniformParcelSize::UniformParcelSize: Quantity that is constant in parcels (base).

flowRateProfile::OpenFOAM.AbstractFlowRateProfile: Type of flow rate profile to apply.

OpenFOAM.PatchInjection — Type

Represents a patch injection model.

name::String: Name to use in injection models list.

patchName::String: Patch where injection is applied.

injectionModel::InjectionModel: Object pointing to general injection model.

U0::Velocity: Particles injection velocity vector.

parcelsPerSecond::Number: Number of parcels injected per second.

sizeDistribution::OpenFOAM.AbstractSizeDistribution: Particle size distribution.

OpenFOAM.TableFlowRateProfile — Type

Table flow rate profile.

OpenFOAM.TabulatedDensity — Type

Tabulated density size distribution.

OpenFOAM.UniformParcelSize — Type

Types of uniform parcel sizes.

OpenFOAM.Velocity — Type

Simple wraper velocity vector.

OpenFOAM.asint — Method

Round and convert number to an integer.

OpenFOAM.parcels_per_second — Method

parcelstoinject2d(;
    mdot::Float64,
    rhop::Float64,
    diam::Float64,
    nParticle::Int64 = 1
)

Computes the flow rate of parcels for a given mean particle size and number of particles per parcels. This is inteded as a helper to create a patchInjection element in the injectionModels of cloudProperties file.

OpenFOAM.spheremass — Method

The mass of a sphere of density rho and diameter diam.

OpenFOAM.spherevolume — Method

The volume of a sphere of diameter diam.

OpenFOAM.tabulate — Method

Compose table entries in typical OpenFOAM format.

Discrete phase models

We can verify the computation of number of parcels per second of sample case injectionChannel. Notice that the value of mdot has been reversed engineered so that it matches the expected value.

julia> parcels_per_second(; ṁ = 0.2, ρ = 1000.0, d = 650.0e-06, nParticle = 1)
1390885